requirements.txt
torch
torchvison
matplotlib
tqdm
numpy
pillow